ISSUE-152 Add PodaacUtils.mine_drive_urls_from_granule_search function

+++ podaac/tests/podaac_test.py
@@ -174,10 +174,10 @@ class TestPodaac(unittest.TestCase):
         path2 = os.path.dirname(__file__)
         self.podaac.granule_subset(input_file_path=path1, path=path2)
 
-        assert os.path.isfile(path2 + 
-            '/subsetted-ascat_20160409_113000_metopa_49153_eps_o_250_2401_ovw.l2.nc') == True
-        os.remove(path2 +
-            '/subsetted-ascat_20160409_113000_metopa_49153_eps_o_250_2401_ovw.l2.nc')
+        if os.path.isfile(path2 +
+                      '/subsetted-ascat_20160409_113000_metopa_49153_eps_o_250_2401_ovw.l2.nc') is True
+            os.remove(path2 +
+                      '/subsetted-ascat_20160409_113000_metopa_49153_eps_o_250_2401_ovw.l2.nc')
 
     # test case for the function subset_status
     def test_subset_status(self):
